IPMI Plugin - Auto Discovery of BMC nodes
With this checkin the IPMI Plugin in aggregator node will automatically
figure out the Compute Nodes' BMC's IP Address under it.

Adding more detail..
Each IPMI plugin in the compute node will read the BMC's IP address
through the KCS driver and send it along with the data blob collected
by the heartbeat plugin. After this the IPMI plugin disables itself(in
compute node only).

The Aggregator's IPMI plugin will collect and unpack this packed
credentials and update a list of BMC nodes it needs to monitor, which
correspond to the compute nodes under its jursidiction.