IPMI Plugin
Added the db store procedure for the statically collected data during
the initial collection point.

Currently only the baseboard serial number is stored, but eventually all
the the static information will be collected and stored in the database
at this point. This is to avoid performance hits while collecting static
information that wouldn't change unless the entire node goes down and
has to be replaced with a new baseboard.